Add files via upload
Global warming is increasing the frequency and intensity of some types of extreme weather. From the analysis, we have gathered useful insights over the past 10 years. for example , warming is causing more rain to fall in heavy downpours. there are also longer dry periods between rainfalls. This, coupled with more evaporation due to higher temperatures. Wet places have become drier. We can observe in a line plot  that the most common weather that prevails is 'Partly Cloudy'. the 'weather v/s temperature' graph shows that the maximum temperature is when the weather is 'Dry'. And the 'weather v/s humidity' graph shows that the maximum humidity is for the weather types- Foggy , Breezy and Foggy, and Rain.